+++ public/userscripts/manga-tracker.user.js
@@ -12,7 +12,7 @@
 // @include      /^https?:\/\/bato\.to\/reader.*$/
 // @include      /^http:/\/dynasty-scans\.com\/chapters\/.+$/
 // @include      /^http:\/\/www\.mangapanda\.com\/(?!(?:search|privacy|latest|alphabetical|popular|random)).+\/.+$/
-// @include      /^https?:\/\/mangastream.com\/r\/.+\/.+\/[0-9]+(?:\/[0-9]+)?$/
+// @include      /^https?:\/\/readms\.net\/r\/.+\/.+\/[0-9]+(?:\/[0-9]+)?(?:\?.+)?$/
 // @include      /^http:\/\/www\.webtoons\.com\/(?:en|zh-hant|zh-hans|th|id)\/[a-z0-9A-Z-_]+\/[a-z0-9A-Z-_]+\/[a-z0-9A-Z-_]+\/viewer\?title_no=[0-9]+&episode_no=[0-9]+$/
 // @include      /^http:\/\/kissmanga\.com\/Manga\/[a-zA-Z0-9-_]+\/[a-zA-Z0-9-_%]+\?id=[0-9]+$/
 // @include      /^https?:\/\/reader\.kireicake\.com\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
@@ -26,8 +26,8 @@
 // @include      /^http:\/\/www\.demonicscans\.com\/FoOlSlide\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
 // @include      /^https?:\/\/reader\.deathtollscans\.net\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
 // @include      /^http:\/\/read\.egscans\.com\/[A-Za-z0-9\-_\!,]+(?:\/Chapter_[0-9]+(?:_extra)?\/?)?$/
-// @updated      2017-02-21
-// @version      1.4.2
+// @updated      2017-03-30
+// @version      1.4.3
 // @downloadURL  https://trackr.moe/userscripts/manga-tracker.user.js
 // @updateURL    https://trackr.moe/userscripts/manga-tracker.meta.js
 // @require      https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js
@@ -967,14 +967,14 @@ let sites = {
 		}
 	}),
 
-	'mangastream.com' : extendSite({
+	'readms.net' : extendSite({ //MangaStream
 		setObjVars : function() {
 			this.page_count  = parseInt($('.controls ul:last > li:last').text().replace(/[^0-9]/g, ''));
 			this.title       = this.segments[2];
 			this.chapter     = this.segments[3]+'/'+this.segments[4];
 
-			this.title_url   = this.https+'://mangastream.com/manga/'+this.title;
-			this.chapter_url = this.https+'://mangastream.com/r/'+this.title+'/'+this.chapter;
+			this.title_url   = this.https+'://readms.net/manga/'+this.title;
+			this.chapter_url = this.https+'://readms.net/r/'+this.title+'/'+this.chapter;
 
 			// this.chapterList     = {}; //This is set via preSetupTopBar.
 			this.chapterListCurrent = this.chapter_url+'/1';
@@ -1003,7 +1003,7 @@ let sites = {
 				},
 				cache: false,
 				success: function(response) {
-					let table = $(response.replace(/^[\S\s]*(<table[\S\s]*<\/table>)[\S\s]*$/, '$1'));
+					let table = $(response.replace(/^[\S\s]*(<table[\S\s]*<\/table>)[\S\s]*$/, '$1').replace(/\?t=[0-9]+&(amp;)?f=[0-9]+&(amp;)?e=[0-9]+/g, ''));
 
 					_this.chapterList = generateChapterList($('tr:not(:first) a', table).reverseObj(), 'href');